pyproject.toml
[project]
name = "fastapi-boot"
version = "0.0.37"
description = "FastAPI tools"
readme = "README.md"
authors = [{ name = "hfdy", email = "hfdy09354121794@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.125.0",
    "pydantic>=2.12.5",
    "pytest-asyncio>=1.3.0",
    "starlette>=0.50.0",
    "uvicorn>=0.38.0",
]


[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = ["tests/test_project"]

[project.scripts]
fastapi-boot = "fastapi_boot.cli.cli:main"

[project.optional-dependencies]
db = ["tortoise-orm>=0.24.1"]
